FP8 training support, fp16 Muon NS, fix GPU→CPU sync stalls
- Add --fp8 flag: torchao Float8Linear for forward/backward matmuls
- Muon Newton-Schulz: bf16 → fp16 (more mantissa precision per Dao-AILab)
- Fix _compute_grad_norm: was .item() per parameter (hundreds of syncs),
  now stacks on GPU and syncs once
- Fix _per_layer_grad_norms: same per-parameter sync issue
- Deduplicate grad_norm.item() in logging (was called 3x, now 1x)
- Canary configs for FP8, micro_batch=8, and profiling runs
- Profile trace parser tool

Canary results (8xB200, 4096 ctx, AttnRes, Liger+compile):
  3B bf16: 228K tok/s  →  3B FP8: 240K tok/s (+5%)
  8B bf16: 111K tok/s  →  8B FP8: 121K tok/s (+9%)

+++ src/training/muon.py
@@ -130,7 +130,7 @@ def newton_schulz_orthogonalize(
 
     coeffs = _resolve_ns_coefficients(ns_coefficients, num_iterations)
     orig_dtype = M.dtype
-    X = M.bfloat16()
+    X = M.half()
 
     # Transpose tall matrices: NS converges faster when ncols >= nrows
     transposed = False
@@ -165,7 +165,7 @@ def _batched_newton_schulz(
     """
     coeffs = _resolve_ns_coefficients(ns_coefficients, num_iterations)
     orig_dtype = M.dtype
-    X = M.bfloat16()
+    X = M.half()
 
     # Transpose if tall
     transposed = False
